cleaned up ReportPage
