+++ oya-web/src/report/ReportPage.tsx
@@ -1,18 +1,17 @@
-import { Box, Grid, Paper, Stack, Tab, Tabs, Typography } from '@mui/material';
+import { Box, Tab, Tabs } from '@mui/material';
 import React from 'react';
 import { Route, Routes, useLocation } from 'react-router';
 import { Link } from 'react-router-dom';
-import { Activity } from '../apiService';
 import BarTotalRechart from './BarTotalRechart';
 import PieTotalRechart from './PieTotalRechart';
-import ReportProvider, { useReportContext } from './ReportProvider';
+import ReportProvider from './ReportProvider';
 import TreemapReactVis from './TreemapReactVis';
 import TreemapRechart from './TreemapRechart';
-import { getDeltaStringOfRange as ts } from '../utils';
 import TreemapNivo from './TreemapNivo';
+import { GridOverviewReport } from './ActivityOverviewReport';
 
 const reportRoutes = [
-  { link: 'cards', label: 'cards', element: <GripOverviewReport /> },
+  { link: 'cards', label: 'cards', element: <GridOverviewReport /> },
   { link: 'bar', label: 'bars', element: <BarTotalRechart /> },
   { link: 'pie', label: 'pie', element: <PieTotalRechart /> },
   { link: 'vis-tree', label: 'vis tree', element: <TreemapReactVis /> },
@@ -43,45 +42,5 @@ export default function ReportPage() {
         </Box>
       </Box>
     </ReportProvider>
-  );
-}
-
-type ActivityOverViewReportProps = {
-  activity: Activity,
-};
-
-function GripOverviewReport() {
-  const { atra } = useReportContext();
-  return (
-    <Grid container spacing={2}>
-      {atra.map((r) => r && (
-        <Grid key={r.activity.id} item xs={6} md={4} lg={3} xl={2}>
-          <ActivityOverViewReport activity={r.activity} />
-        </Grid>
-      ))}
-    </Grid>
-  );
-}
-
-export function ActivityOverViewReport({ activity }: ActivityOverViewReportProps) {
-  const { atrm } = useReportContext();
-  const renderRow = (name: string, value: string | number) => (
-    <Typography textAlign="center" key={name}>
-      {name}: <b>{value}</b>
-    </Typography>
-  );
-  const r = atrm[activity.id];
-  if (!r) return null;
-  return (
-    <Paper sx={{ p: 2, height: '100%', flexShrink: 0 }}>
-      <Typography gutterBottom variant='h5' textAlign="center">{activity.name}</Typography>
-      <Stack>
-        {renderRow('days', `${r.days} of ${r.allDays}`)}
-        {renderRow('total time', ts(r.time))}
-        {renderRow('avg per all days', ts(r.avgPerAllDays))}
-        {renderRow('avg per days', ts(r.avgPerDays))}
-        {renderRow('occurance', r.occurance.toString())}
-      </Stack>
-    </Paper>
   );
 }
